Plant-Disease-Detection-Keras-FastAPI/requirements.txt
fastapi[all]
Keras-2.9.0
tensorflow-2.9.1
numpy-1.23.1
pillow-9.2.0
opencv_python==4.4.0.46
